# search: skip timing the unoptimized kernel