Also:

* ensure the return the unoptimized kernel if no opts are valid
* refactor debugging to a single BEAM_DEBUG variable.
* added an early stop as well if the best kernel runs in under `min_progress` since the next iteration cannot improve by more than that.

for tiny red / `HSA=1`:

| branch  | settings | setup (m:ss) | step time (ms) | flops (TF) |
| --- | --- | --- | --- |  --- |
| master | default | 26:24 | 365ms | 104TF |
| search_speed_improvements  | default | 23:56 | 367ms | 103TF |

for tiny green / `CUDA=1`:

| branch  | settings | setup (m:ss) | step time (ms) | flops (TF) |
| --- | --- | --- | --- |  --- |
| master | default | 23:53 | 263ms | 144TF |
| search_speed_improvements | default | 19:53 | 263ms | 144TF |